RoundsVersion2/easyocr_server/requirements.txt
easyocr>=1.7.0
flask>=2.3.0
flask-cors==4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-image>=0.21.0
scipy>=1.11.0
requests>=2.31.0
gunicorn==21.2.0 